LPS-74542 Remove references to dynamic because now there's only one behaviour
