+++ adaptive-media-blogs-editor-configuration/bnd.bnd
@@ -1,5 +1,5 @@
-Bundle-Name: Liferay Adaptive Media Blogs Editor Configuration Dynamic
-Bundle-SymbolicName: com.liferay.adaptive.media.blogs.editor.configuration.dynamic
+Bundle-Name: Liferay Adaptive Media Blogs Editor Configuration
+Bundle-SymbolicName: com.liferay.adaptive.media.blogs.editor.configuration
 Bundle-Version: 1.0.0
 Liferay-Releng-Module-Group-Description:
 Liferay-Releng-Module-Group-Title: Adaptive Media
